Implement sending email functionality
We need to have valid email addresses of our users in order to be able to send them emails in the future. In this patch, our system should send email after an account is created from info@studyagenda.com with activation link. When student clicks the link, it activates his account and he can start using the system. Also, the "forgot password?" functionality should be implemented with sending an email.

Fixes #40