Improve the "Add Task" process
In the "Add Task" modal, ask first for course, then due date. If there is already a task that matches moth course and due date, populate task name and weight. If student save an existing task, and it already exists in student's schedule the system should not create a new task.

Fixes #57